+++ src/api/rpc/full_node/index.ts
@@ -253,6 +253,21 @@ export async function get_coin_record_by_name(agent: TRPCAgent, data: TGetCoinRe
 
 
 
+
+export const get_coin_records_by_parent_ids_command = "get_coin_records_by_parent_ids";
+export type get_coin_records_by_parent_ids_command = typeof get_coin_records_by_parent_ids_command;
+export type TGetCoinRecordsByParentIdsRequest = {
+  parent_ids: str[];
+  start_height?: uint32;
+  end_height?: uint32;
+  include_spent_coins?: bool;
+};
+export type TGetCoinRecordsByParentIdsResponse = {
+  coin_records: CoinRecord[];
+};
+
+
+
 export const push_tx_command = "push_tx";
 export type push_tx_command = typeof push_tx_command;
 export type TPushTxRequest = {